backend/ReportAgent/requirements.txt
# Django and REST framework
Django>=4.2.0,<5.0.0
djangorestframework>=3.14.0
django-cors-headers>=4.3.0

# Async and WebSocket support
channels>=4.0.0
channels-redis>=4.1.0
daphne>=4.0.0

# Task Queue and Cache
celery>=5.3.0
redis>=5.0.0
flower>=2.0.0

# Machine Learning and Image Processing
torch>=2.0.0
torchvision>=0.15.0
transformers>=4.35.0
segment-anything @ git+https://github.com/facebookresearch/segment-anything.git
accelerate
timm
Pillow>=10.0.0
numpy>=1.24.0
rembg>=2.0.0
opencv-python>=4.8.0
scikit-image>=0.21.0
onnxruntime>=1.15.1

# Web3 and Blockchain
web3>=6.11.1
eth-account>=0.8.0
eth-typing>=3.4.0
eth-utils>=2.2.0

# File Processing
python-magic>=0.4.27
python-dotenv>=1.0.0
pyyaml>=6.0.1

# Utilities
tqdm>=4.65.0
requests>=2.31.0
urllib3>=2.0.0
python-dateutil>=2.8.2

# Development & Debugging
ipython>=8.0.0
pytest>=7.4.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0

# WebSocket Testing
websockets>=11.0.0

# Security
python-jose>=3.3.0
cryptography>=41.0.0

# Production
gunicorn>=21.2.0
uvicorn>=0.23.0